+++ modules/frontend/combiner/search_tag_values.go
@@ -4,6 +4,7 @@ import (
 	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/collector"
 	"github.com/grafana/tempo/pkg/tempopb"
+	"go.uber.org/atomic"
 )
 
 var (
@@ -14,6 +15,7 @@ var (
 func NewSearchTagValues(limitBytes int) Combiner {
 	// Distinct collector with no limit
 	d := collector.NewDistinctStringWithDiff(limitBytes)
+	inspectedBytes := atomic.NewUint64(0)
 
 	c := &genericCombiner[*tempopb.SearchTagValuesResponse]{
 		httpStatusCode: 200,
@@ -23,10 +25,16 @@ func NewSearchTagValues(limitBytes int) Combiner {
 			for _, v := range partial.TagValues {
 				d.Collect(v)
 			}
+			if partial.Metrics != nil {
+				inspectedBytes.Add(partial.Metrics.InspectedBytes)
+			}
 			return nil
 		},
 		finalize: func(final *tempopb.SearchTagValuesResponse) (*tempopb.SearchTagValuesResponse, error) {
 			final.TagValues = d.Strings()
+			// return metrics in final response
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			final.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return final, nil
 		},
 		quit: func(_ *tempopb.SearchTagValuesResponse) bool {
@@ -38,6 +46,9 @@ func NewSearchTagValues(limitBytes int) Combiner {
 				return nil, err
 			}
 			response.TagValues = resp
+			// also return latest metrics along with diff
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			response.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return response, nil
 		},
 	}
@@ -52,6 +63,7 @@ func NewTypedSearchTagValues(limitBytes int) GRPCCombiner[*tempopb.SearchTagValu
 func NewSearchTagValuesV2(limitBytes int) Combiner {
 	// Distinct collector with no limit and diff enabled
 	d := collector.NewDistinctValueWithDiff(limitBytes, func(tv tempopb.TagValue) int { return len(tv.Type) + len(tv.Value) })
+	inspectedBytes := atomic.NewUint64(0)
 
 	c := &genericCombiner[*tempopb.SearchTagValuesV2Response]{
 		httpStatusCode: 200,
@@ -60,6 +72,9 @@ func NewSearchTagValuesV2(limitBytes int) Combiner {
 		combine: func(partial, _ *tempopb.SearchTagValuesV2Response, _ PipelineResponse) error {
 			for _, v := range partial.TagValues {
 				d.Collect(*v)
+			}
+			if partial.Metrics != nil {
+				inspectedBytes.Add(partial.Metrics.InspectedBytes)
 			}
 			return nil
 		},
@@ -70,6 +85,9 @@ func NewSearchTagValuesV2(limitBytes int) Combiner {
 				v2 := v
 				final.TagValues = append(final.TagValues, &v2)
 			}
+			// load Inspected Bytes here and return along with final response
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			final.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return final, nil
 		},
 		quit: func(_ *tempopb.SearchTagValuesV2Response) bool {
@@ -85,6 +103,9 @@ func NewSearchTagValuesV2(limitBytes int) Combiner {
 				v2 := v
 				response.TagValues = append(response.TagValues, &v2)
 			}
+			// also return metrics along with diffs
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			response.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return response, nil
 		},
 	}