+++ modules/frontend/combiner/search_tags.go
@@ -4,6 +4,7 @@ import (
 	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/collector"
 	"github.com/grafana/tempo/pkg/tempopb"
+	"go.uber.org/atomic"
 )
 
 var (
@@ -13,6 +14,7 @@ var (
 
 func NewSearchTags(limitBytes int) Combiner {
 	d := collector.NewDistinctStringWithDiff(limitBytes)
+	inspectedBytes := atomic.NewUint64(0)
 
 	c := &genericCombiner[*tempopb.SearchTagsResponse]{
 		httpStatusCode: 200,
@@ -22,10 +24,16 @@ func NewSearchTags(limitBytes int) Combiner {
 			for _, v := range partial.TagNames {
 				d.Collect(v)
 			}
+			if partial.Metrics != nil {
+				inspectedBytes.Add(partial.Metrics.InspectedBytes)
+			}
 			return nil
 		},
 		finalize: func(response *tempopb.SearchTagsResponse) (*tempopb.SearchTagsResponse, error) {
 			response.TagNames = d.Strings()
+			// return metrics with final results
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			response.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return response, nil
 		},
 		quit: func(_ *tempopb.SearchTagsResponse) bool {
@@ -38,6 +46,9 @@ func NewSearchTags(limitBytes int) Combiner {
 			}
 
 			response.TagNames = resp
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			// return metrics with diff results
+			response.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return response, nil
 		},
 	}
@@ -52,6 +63,7 @@ func NewTypedSearchTags(limitBytes int) GRPCCombiner[*tempopb.SearchTagsResponse
 func NewSearchTagsV2(limitBytes int) Combiner {
 	// Distinct collector map to collect scopes and scope values
 	distinctValues := collector.NewScopedDistinctStringWithDiff(limitBytes)
+	inspectedBytes := atomic.NewUint64(0)
 
 	c := &genericCombiner[*tempopb.SearchTagsV2Response]{
 		httpStatusCode: 200,
@@ -62,6 +74,9 @@ func NewSearchTagsV2(limitBytes int) Combiner {
 				for _, tag := range res.Tags {
 					distinctValues.Collect(res.Name, tag)
 				}
+			}
+			if partial.Metrics != nil {
+				inspectedBytes.Add(partial.Metrics.InspectedBytes)
 			}
 			return nil
 		},
@@ -75,6 +90,9 @@ func NewSearchTagsV2(limitBytes int) Combiner {
 					Tags: vals,
 				})
 			}
+			// return metrics with final results
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			final.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return final, nil
 		},
 		quit: func(_ *tempopb.SearchTagsV2Response) bool {
@@ -93,7 +111,9 @@ func NewSearchTagsV2(limitBytes int) Combiner {
 					Tags: vals,
 				})
 			}
-
+			// TODO: merge with other metrics as well, when we have them, return only InspectedBytes for now
+			// also return metrics with diff results
+			response.Metrics = &tempopb.MetadataMetrics{InspectedBytes: inspectedBytes.Load()}
 			return response, nil
 		},
 	}